Refactor files Test
