+++ get-that-job/src/components/Home/Home.js
@@ -1,14 +1,14 @@
 import colors from "../../styles/colors";
 import styled from "@emotion/styled";
-import { StyledCreateAccountButton } from "../buttons/buttons";
-import group from "../img/group.svg";
-import searchLogo from "../img/searchLogo.svg";
-import dik from "../img/dik.svg";
-import fatima from "../img/fatima.svg";
-import felix from "../img/felix.svg";
-import jorge from "../img/jorge.svg";
-import josebe from "../img/josebe.svg";
-import joset from "../img/joset.svg";
+import { StyledCreateAccountButton } from "../Buttons/Buttons";
+import group from "../Img/group.svg";
+import searchLogo from "../Img/searchLogo.svg";
+import dik from "../Img/dik.svg";
+import fatima from "../Img/fatima.svg";
+import felix from "../Img/felix.svg";
+import jorge from "../Img/jorge.svg";
+import josebe from "../Img/josebe.svg";
+import joset from "../Img/joset.svg";
 import git from "../pic/git.svg";
 import linkedin from "../pic/linkedin.svg";
 import ruby from "../pic/ruby.svg";
