--- get-that-job/src/components/buttons/buttons.js
@@ -1,216 +1,0 @@
-import colors from "../../styles/colors";
-import styled from "@emotion/styled";
-
-export const StyledCreateAccountButton = styled.button`
-  height: 56px;
-  width: 264px;
-  font-size: 14px;
-  font-weight: 500;
-  border-radius: 16px;
-  background-color: ${colors.pink};
-  color: ${colors.white};
-  border: none;
-`;
-
-export const StyledLogInButton = styled.button`
-  height: 40px;
-  width: 80px;
-  left: 276px;
-  top: 136px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.pink};
-`;
-
-export const StyledNextButton = styled.button`
-  display: flex;
-  height: 40px;
-  width: 106px;
-  left: 127px;
-  top: 204px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.pink};
-  gap: 8px;
-`;
-
-export const StyledSkipButton = styled.button`
-  height: 40px;
-  width: 115px;
-  left: 0px;
-  top: 0px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.background};
-  border-color: ${colors.pink};
-`;
-
-export const StyledChooseFileButton = styled.button`
-  display: flex;
-  height: 36px;
-  width: 134px;
-  border-radius: 8px;
-  padding: 8px;
-  background-color: ${colors.pink};
-  gap: 8px;
-`;
-
-export const StyledPreviousButton = styled.button`
-  display: flex;
-  width: 131px;
-  height: 40px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  gap: 8px;
-`;
-
-export const StyledCurrentNumber = styled.div`
-  width: 32px;
-  height: 32px;
-  border-radius: 66.6667px;
-  background-color: ${colors.pink};
-`;
-
-export const StyledNotCurrentNumber = styled.div`
-  width: 32px;
-  height: 32px;
-  border-radius: 66.6667px;
-  background-color: ${colors.gray};
-`;
-
-export const StyledRadioButton = styled.button`
-  height: 20px;
-  width: 20px;
-  border-radius: 50px;
-`;
-
-export const StyledCurrentRadioButton = styled.div`
-  height: 12px;
-  width: 12px;
-  left: 4px;
-  top: 4px;
-  border-radius: 50px;
-  background-color: ${colors.pink};
-`;
-
-export const StyledSeeMoreButton = styled.button`
-  height: 40px;
-  width: 112px;
-  left: 146px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  border-color: ${colors.pink};
-  display: flex;
-  align-items: center;
-  background-color: ${colors.white};
-`;
-
-export const StyledFollowButton = styled.button`
-  border: none;
-  display: flex;
-  align-items: center;
-  gap: 4px;
-  font-size: 14px;
-  color: ${colors.gray};
-  background-color: ${colors.white};
-  gap: 13px;
-  padding: 16px;
-`;
-
-export const StyledApplyNowButton = styled.button`
-  height: 56px;
-  width: 173px;
-  top: 12px;
-  border-radius: 16px;
-  padding: 16px 24px 16px 24px;
-  background-color: ${colors.pink};
-  display: flex;
-  align-items: center;
-`;
-
-export const StyledCloseButton = styled.button`
-  height: 40px;
-  width: 116px;
-  left: 134px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.darkPink};
-  display: flex;
-  align-items: center;
-  gap: 8px;
-`;
-
-export const StyledShowButton = styled.button`
-  border: none;
-  display: flex;
-  align-items: center;
-  gap: 8px;
-  padding: 8px 16px;
-  background-color: ${colors.white};
-`;
-export const StyledMarkStartedButton = styled.button`
-  height: 40px;
-  width: 180px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  border-color: ${colors.pink};
-  align-items: center;
-`;
-
-export const StyledDownLoadButton = styled.button`
-  height: 40px;
-  width: 183px;
-  left: 288.5px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.white};
-  border-color: ${colors.pink};
-  display: flex;
-  align-items: center;
-`;
-
-export const StyledUpdateProfileButton = styled.button`
-  height: 40px;
-  width: 165px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.pink};
-  align-items: center;
-`;
-export const StyledDeclineButton = styled.button`
-  height: 40px;
-  width: 242px;
-  left: 422.5px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  background-color: ${colors.darkPink};
-  display: flex;
-  align-items: center;
-  gap: 8px;
-`;
-export const StyledPostJobButton = styled.button`
-  height: 40px;
-  width: 153px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  align-items: center;
-  background-color: ${colors.pink};
-`;
-
-export const StyledSaveButton = styled.button`
-  height: 40px;
-  width: 155px;
-  border-radius: 16px;
-  padding: 8px 16px 8px 16px;
-  align-items: center;
-  background-color: ${colors.pink};
-`;
-
-// export const StyledShowButton = styled.button`
-//   height: 56px;
-//   width: 233px;
-//   border-radius: 16px;
-//   padding: 16px 24px 16px 24px;
-//   align-items: center;
-//   background-color: ${colors.pink};
-// `;
