--- get-that-job/src/components/forms/inputForm.js
@@ -1,12 +1,0 @@
-import colors from "../../styles/colors";
-import styled from "@emotion/styled";
-
-const StyledInput = styled.input`
-  background: ${colors.gray};
-`;
-
-function InputForm() {
-  return <StyledInput />;
-}
-
-export default InputForm;
